requirements.txt
gevent==1.0.1
gevent-psycopg2==0.0.3
greenlet==0.4.3
gunicorn==19.1.1
itsdangerous==0.24
psycopg2==2.5.3
python-memcached==1.53
wsgiref==0.1.2
